feat: Moved view to right folder
