+++ apps/dashboard/src/modules/admin/views/AdminRBACView.vue
